[CodeMenu] Installed Player Slot Color Changer
Modified and replaced all the files necessary to install the Player Slot Color Changer. Rather importantly, this makes use of an adaptation of the code to allow the code to be externally linked into the codeset. If any issues come up with this, please let me know, this is a maiden voyage for this kind of thing lol.

Additionally, disabled "P3 on results screen is purple [Bird]" in BOOST.txt, as the Color Changer code handles the colors of names on the Results Screen independently.